+++ blog-springboot/src/main/java/com/wzh/blog/service/impl/CategoryServiceImpl.java
@@ -15,7 +15,7 @@ import com.wzh.blog.entity.Category;
 import com.wzh.blog.dao.CategoryDao;
 import com.wzh.blog.exception.BizException;
 import com.wzh.blog.service.CategoryService;
-import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
+import com.baomidou.mybatisplus.spring.service.impl.ServiceImpl;
 import com.wzh.blog.vo.CategoryVO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -37,15 +37,20 @@ public class CategoryServiceImpl extends ServiceImpl<CategoryDao, Category> impl
     @Autowired
     private ArticleDao articleDao;
 
+
+
+
     @Override
     public PageResult<CategoryDTO> listCategories() {
         return new PageResult<>(categoryDao.listCategoryDTO(), categoryDao.selectCount(null));
     }
 
+
+
     @Override
     public PageResult<CategoryBackDTO> listBackCategories(ConditionVO condition) {
         // 查询分类数量
-        Integer count = categoryDao.selectCount(new LambdaQueryWrapper<Category>()
+        Long count = categoryDao.selectCount(new LambdaQueryWrapper<Category>()
                 .like(StringUtils.isNotBlank(condition.getKeywords()), Category::getCategoryName, condition.getKeywords()));
         if (count == 0) {
             return new PageResult<>();
@@ -54,6 +59,8 @@ public class CategoryServiceImpl extends ServiceImpl<CategoryDao, Category> impl
         List<CategoryBackDTO> categoryList = categoryDao.listCategoryBackDTO(PageUtils.getLimitCurrent(), PageUtils.getSize(), condition);
         return new PageResult<>(categoryList, count);
     }
+
+
 
     @Override
     public List<CategoryOptionDTO> listCategoriesBySearch(ConditionVO condition) {
@@ -64,16 +71,20 @@ public class CategoryServiceImpl extends ServiceImpl<CategoryDao, Category> impl
         return BeanCopyUtils.copyList(categoryList, CategoryOptionDTO.class);
     }
 
+
+
     @Override
     public void deleteCategory(List<Integer> categoryIdList) {
         // 查询分类id下是否有文章
-        Integer count = articleDao.selectCount(new LambdaQueryWrapper<Article>()
+        Long count = articleDao.selectCount(new LambdaQueryWrapper<Article>()
                 .in(Article::getCategoryId, categoryIdList));
         if (count > 0) {
             throw new BizException("删除失败，该分类下存在文章");
         }
         categoryDao.deleteBatchIds(categoryIdList);
     }
+
+
 
     @Override
     public void saveOrUpdateCategory(CategoryVO categoryVO) {
